# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ShareDirStat API
  version: "1"
  summary: Disk usage analysis for mounted shares, with delete and download.
  description: |
    The JSON API behind the ShareDirStat web UI. Every endpoint is also
    intended for scripts.

    **No authentication.** Access control is the job of the surrounding
    infrastructure (reverse proxy, VPN, ingress middleware).

    **Mutating requests** (`POST`) must carry the header
    `X-Requested-With: ShareDirStat` and a JSON body. Cross-origin browser
    requests are refused: no CORS headers are ever emitted, and a
    `Sec-Fetch-Site` other than `same-origin`/`none` is rejected.

    **Paths** are share-relative, `/`-separated, without a leading slash; the
    empty string is the share root. File names on Linux are bytes, not text:
    a name that is not valid UTF-8 is returned with replacement characters in
    `name` and its raw bytes base64-encoded in `name_b64`.

    **Sizes** are bytes. Each node carries both `size` (apparent, `st_size`)
    and `alloc` (allocated, `st_blocks * 512`); the `basis` query parameter
    selects which one governs sorting and percentages.

    **Caching.** Tree, treemap, top, extension and search responses carry an
    `ETag` derived from the share's current generation and mutation counter;
    `If-None-Match` returns `304`.
  license:
    name: MIT
    url: https://github.com/StevenGann/ShareDirStat/blob/main/LICENSE
servers:
  - url: /api/v1
...
paths:
...
  /shares/{id}/scans:
    get:
      tags: [scans]
      summary: Scan history
      operationId: scanHistory
      parameters: [{ $ref: "#/components/parameters/shareId" }]
      responses:
        "200":
          description: Past scans newest first, plus the running one if any.
          content:
            application/json:
              schema:
                type: object
                properties:
                  scans:
                    type: array
                    items: { $ref: "#/components/schemas/ScanRecord" }
                  running:
                    oneOf:
                      - { $ref: "#/components/schemas/ScanStatus" }
                      - { type: "null" }
  /shares/{id}/scan:
    post:
      tags: [scans]
      summary: Start a scan
      description: |
        With an empty or absent `path`, scans the whole share into a fresh
        result set; the previous results stay served until it completes. With
        a `path`, rescans only that subtree and splices it into the current
        results.
      operationId: startScan
      parameters: [{ $ref: "#/components/parameters/shareId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                path: { type: string, default: "" }
      responses:
        "202":
          description: Accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanStatus" }
        "403": { $ref: "#/components/responses/CSRF" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: A scan is already running, or the share is unavailable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /shares/{id}/download:
    get:
      tags: [files]
      summary: Download one file
      description: |
        Streams a regular file with `Content-Disposition: attachment`. Range
        requests are honoured (`206`), so a large download can be resumed.
        Media types a browser might execute (`text/html`, `image/svg+xml`,
        JavaScript, XML) are served as `application/octet-stream`.

        This endpoint reads the filesystem directly, so a file created after
        the last scan is still downloadable.
      operationId: download
      parameters:
        - { $ref: "#/components/parameters/shareId" }
        - { $ref: "#/components/parameters/path" }
      responses:
        "200": { description: The file. }
        "206": { description: The requested range. }
        "400":
          description: Not a regular file, the share root, or an unsafe path.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { description: Downloading is not enabled for this share. }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /shares/{id}/trash:
    get:
      tags: [files]
      summary: Items in the trash
      operationId: listTrash
      parameters: [{ $ref: "#/components/parameters/shareId" }]
      responses:
        "200":
          description: Newest first.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/TrashItem" }
        "403": { description: Trash mode is not enabled. }
  /shares/{id}/trash/restore:
    post:
      tags: [files]
      summary: Restore an item from the trash
      description: Moves the item back to its original location and queues a rescan of that folder.
      operationId: restoreTrash
      parameters: [{ $ref: "#/components/parameters/shareId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string, description: The item's `trash_path`. }
      responses:
        "200":
          description: Restored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TrashItem" }
        "409": { description: Something already exists at the original location. }
  /shares/{id}/trash/empty:
    post:
      tags: [files]
      summary: Empty the trash
      operationId: emptyTrash
      parameters: [{ $ref: "#/components/parameters/shareId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                batch:
                  type: string
                  description: One batch's timestamp name; empty means everything.
      responses:
        "200":
          description: Done.
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed: { type: integer }
...
components:
  parameters:
    shareId:
      name: id
      in: path
      required: true
      schema: { type: string, pattern: "^[a-z0-9][a-z0-9_-]{0,63}$" }
...
    path:
      name: path
      in: query
      description: Share-relative path; empty for the root.
      schema: { type: string, default: "" }
...
  responses:
...
    NotFound:
      description: No such share, scan or path.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    CSRF:
      description: Missing `X-Requested-With` header or cross-origin request (`csrf_rejected`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: A stable machine-readable code.
              examples: [share_not_found, path_not_found, csrf_rejected, confirmation_required, delete_disabled]
            message: { type: string }
            details: { type: object }
...
    ScanStatus:
      type: object
      properties:
        id: { type: string }
        share_id: { type: string }
        trigger: { type: string, enum: [manual, schedule, startup, rescan, reconcile] }
        path: { type: string, description: Empty for a full scan. }
        started_at: { type: string, format: date-time }
        queued: { type: boolean }
        paused: { type: boolean }
        progress:
          type: object
          properties:
            dirs: { type: integer }
            files: { type: integer }
            bytes: { type: integer }
            errors: { type: integer }
            excluded: { type: integer }
            pending: { type: integer }
            current_path: { type: string }
            elapsed_ms: { type: integer }
            rate_files_per_s: { type: number }
            paused: { type: boolean }
    ScanRecord:
      type: object
      properties:
        id: { type: string }
        share_id: { type: string }
        trigger: { type: string }
        path: { type: string }
        started_at: { type: string, format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        duration_ms: { type: integer }
        outcome: { type: string, enum: [completed, cancelled, failed] }
        error: { type: string }
        generation: { type: string }
        files: { type: integer }
        dirs: { type: integer }
        bytes: { type: integer }
        errors: { type: integer }
...
    TrashItem:
      type: object
      properties:
        batch: { type: string }
        trash_path: { type: string }
        original_path: { type: string }
        deleted_at: { type: string, format: date-time }
        kind: { type: string }
        size: { type: integer }
        blocked: { type: boolean, description: The original location is occupied again. }
...